Deleting of users + constants for URLS are created

+++ src/test/java/site/nomoreparties/stellarburgers/CreateUserTest.java
@@ -1,54 +1,74 @@
 package site.nomoreparties.stellarburgers;
 
-import io.qameta.allure.Issue;
 import io.qameta.allure.junit4.DisplayName;
+import org.junit.After;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
 import site.nomoreparties.stellarburgers.pojo.Client.ApiClient;
 import site.nomoreparties.stellarburgers.pojo.Model.UserData;
-
 import java.util.Random;
 
 public class CreateUserTest {
+    private String email;
+    private String name;
+    private String password;
     private ApiClient client;
+    private UserData userData;
+    private UserData responseRegistration;
 
     @Before
     public void setUp() {
         client = new ApiClient();
+        //сгенерировать данные для регистрации
+        email = new Random().nextInt(100) + "@mail.ru";
+        name = new Random().nextInt(100) + "Mia";
+        password = "12345";
+        //объявить данные пользователя
+        userData = new UserData(email, password, name, null, null, null, null, null);
+    }
+
+    @After
+    public void deleteUser() {
+        //если пользователь был создан
+        if(responseRegistration.getAccessToken() != null) {
+            //получить токен пользователя, сгенерированный при регистрации
+            String accessToken = responseRegistration.getAccessToken();
+            //удалить пользователя
+            client.deleteUser(accessToken);
+        }
     }
 
     @Test
     @DisplayName("Создать уникального пользователя")
     public void validCreateUserTest() {
-        final UserData registration = new UserData("a" + new Random().nextInt(1000) + "@ya.ru", "12345", "Anna", null, null, null, null, null);
-        UserData responseRegistration = client.createRegistration(registration)
+        //зарегистрировать юзера
+        responseRegistration = client.createRegistration(userData)
                 .then()
                 .statusCode(200)
                 .extract().as(UserData.class);
         //проверить, что вернулся не пустой ответ
         Assert.assertNotNull(responseRegistration);
-
-        String accessToken = responseRegistration.getAccessToken();
-        client.deleteUser(accessToken);
     }
 
     @Test
     @DisplayName("Создать пользователя, который уже зарегистрирован")
     public void createNotUniqueUserTest() {
-        final UserData registration = new UserData("test-data@yandex.ru", "12345", "Anna", null, null, null, null, null);
-        final UserData responseRegistration = client.createRegistration(registration)
+        //зарегистрировать юзера
+        responseRegistration = client.createRegistration(userData).then().statusCode(200).extract().as(UserData.class);
+        //зарегистрировать этого же пользователя еще раз
+        UserData responseRegistrationNotUniqueUser = client.createRegistration(userData)
                 .then()
                 .statusCode(403)
                 .extract().as(UserData.class);
-        Assert.assertEquals("User already exists", responseRegistration.getMessage());
+        Assert.assertEquals("User already exists", responseRegistrationNotUniqueUser.getMessage());
     }
 
     @Test
     @DisplayName("Создать пользователя и не заполнить одно из обязательных полей: емейл")
     public void createUserWithEmptyEmailFieldTest() {
-        final UserData registration = new UserData("", "12345", "Anna", null, null, null, null, null);
-        final UserData responseRegistration = client.createRegistration(registration)
+        userData = new UserData("", password, name, null, null, null, null, null);
+        responseRegistration = client.createRegistration(userData)
                 .then()
                 .statusCode(403)
                 .extract().as(UserData.class);
@@ -58,13 +78,11 @@ public class CreateUserTest {
     @Test
     @DisplayName("Создать пользователя и не заполнить одно из обязательных полей: пароль")
     public void createUserWithEmptyPasswordFieldTest() {
-        final UserData registration = new UserData("a" + new Random().nextInt(500) + "@ya.ru", "", "Anna", null, null, null, null, null);
-        final UserData responseRegistration = client.createRegistration(registration)
+        userData = new UserData(email, "", name, null, null, null, null, null);
+        responseRegistration = client.createRegistration(userData)
                 .then()
                 .statusCode(403)
                 .extract().as(UserData.class);
         Assert.assertEquals("Email, password and name are required fields", responseRegistration.getMessage());
     }
-
-
 }
