+++ src/test/java/site/nomoreparties/stellarburgers/UpdateUserDataWithoutAuthorizationTest.java
@@ -13,6 +13,7 @@ import java.util.Random;
 
 public class UpdateUserDataWithoutAuthorizationTest {
     private ApiClient client;
+    private UserData responseRegistration;
     private String accessToken;
 
     @Before
@@ -25,13 +26,13 @@ public class UpdateUserDataWithoutAuthorizationTest {
     @Issue("BUG-2")
     public void updateUserDataWithoutAuthorizationTest() {
         //объявить имя и почту для регистрации
-        String email = "a" + new Random().nextInt(500) + "@ya.ru";
+        String email = new Random().nextInt(1000) + "@ya.ru";
         String name = "Anna";
         String password = "12345";
         //создать юзера
         final UserData userData = new UserData(email, password, name, null, null, null, null, null);
         //зарегистрировать юзера
-        final UserData responseRegistration = client.createRegistration(userData)
+        responseRegistration = client.createRegistration(userData)
                 .then()
                 .extract().as(UserData.class);
         //получить аксесс токен, сгенерированный при регистрации
@@ -45,17 +46,19 @@ public class UpdateUserDataWithoutAuthorizationTest {
         //объявить новые данные пользователя
         final UserData newUserData = new UserData(new Random().nextInt(500) + email, password, name + new Random().nextInt(50), null, null, null, null, null);
         //отправить на сервер новые данные
-        final UserData updatedData = client.updateUserData(accessToken, newUserData)
+        final UserData responseUpdatedData = client.updateUserData(accessToken, newUserData)
                 .then()
                 .statusCode(401)
                 .extract().as(UserData.class);
         //Баг: должна быть ошибка, так как мы обновляем данные пользователя без авторизации, но возвращается 200
-        Assert.assertFalse(updatedData.getSuccess());
-        Assert.assertEquals("You should be authorised", updatedData.getMessage());
+        Assert.assertFalse(responseUpdatedData.getSuccess());
+        Assert.assertEquals("You should be authorised", responseUpdatedData.getMessage());
     }
 
     @After
     public void deleteUser() {
-        client.deleteUser(accessToken);
+        if(responseRegistration != null){
+            client.deleteUser(accessToken);
+        }
     }
 }
